added a web precense template with links to all of my things on the web
i want to do more development on this site, but i cant with my current
computer situation. i cant develop on my current laptop because i need
xcode but it's too old. i need the python developer tools... maybe ill
get a newer mac or a cheap pc and convert it to linux...