another huge commit... but mostly good new and good progress
got the admin panel up and running with the quotes section actually
pulling data down from the admin thing to be used in a view and
a template

to do:
for some reason uploading an image throws phat errors. i think it
might be related to PIL and image libraries on a mac. look here:
http://djangodays.com/2008/09/03/django-imagefield-validation-error-caused-by-incorrect-pil-installation-on-mac/
Still need to work on the javascript nav bar and stripping away
all spacescout stuff. look in the form section.